refactor(test): test-support에서 게임 도메인 의존 설정 제거
test-support는 순수 테스트 인프라(컨테이너·스케줄러·STOMP 헬퍼)만 제공하도록 정리.
게임 스케줄러 mock 설정은 app 모듈이 전담하도록 이동.

- test-support 삭제: BaseServiceTestConfig, BaseIntegrationTestConfig,
  WebTestConfig, ServiceTest, IntegrationTestSupport, @IntegrationTest
- app 복원: global/ServiceTest, global/config/ServiceTestConfig
- 영향 파일 import 재조정(25개)
