+++ backend/app/src/test/java/coffeeshout/report/application/ReportFacadeMemberTest.java
@@ -2,7 +2,7 @@ package coffeeshout.report.application;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import coffeeshout.support.ServiceTest;
+import coffeeshout.global.ServiceTest;
 import coffeeshout.minigame.domain.MiniGameType;
 import coffeeshout.report.domain.ReportCategory;
 import coffeeshout.report.infra.persistence.Report;
