+++ backend/app/src/test/java/coffeeshout/user/application/service/AuthTokenServiceTest.java
@@ -4,7 +4,7 @@ import static coffeeshout.support.ExceptionAssertions.assertCoffeeShoutException
 import static org.assertj.core.api.Assertions.assertThat;
 
 import coffeeshout.fixture.UserFixture;
-import coffeeshout.support.ServiceTest;
+import coffeeshout.global.ServiceTest;
 import coffeeshout.user.domain.TokenPair;
 import coffeeshout.user.domain.AuthenticatedUser;
 import coffeeshout.user.domain.User;
