+++ backend/app/src/test/java/coffeeshout/user/application/service/TermsServiceTest.java
@@ -3,7 +3,7 @@ package coffeeshout.user.application.service;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import coffeeshout.fixture.UserFixture;
-import coffeeshout.support.ServiceTest;
+import coffeeshout.global.ServiceTest;
 import coffeeshout.user.domain.User;
 import coffeeshout.user.domain.repository.UserRepository;
 import coffeeshout.user.infra.persistence.UserEntity;
