+++ backend/app/src/test/java/coffeeshout/user/application/service/UserWithdrawalServiceTransactionTest.java
@@ -4,7 +4,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import coffeeshout.support.TestContainerSupport;
 import coffeeshout.fixture.UserFixture;
-import coffeeshout.support.config.BaseServiceTestConfig;
+import coffeeshout.global.config.ServiceTestConfig;
 import coffeeshout.user.domain.User;
 import coffeeshout.user.domain.repository.RefreshTokenRepository;
 import coffeeshout.user.domain.repository.UserRepository;
@@ -17,7 +17,7 @@ import org.springframework.context.annotation.Import;
 import org.springframework.test.context.ActiveProfiles;
 
 @SpringBootTest
-@Import(BaseServiceTestConfig.class)
+@Import(ServiceTestConfig.class)
 @ActiveProfiles("test")
 class UserWithdrawalServiceTransactionTest extends TestContainerSupport {
