--- backend/test-support/src/main/java/coffeeshout/support/annotation/IntegrationTest.java
@@ -1,21 +1,0 @@
-package coffeeshout.support.annotation;
-
-import coffeeshout.support.config.BaseIntegrationTestConfig;
-import coffeeshout.support.config.WebTestConfig;
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.context.annotation.Import;
-import org.springframework.test.context.ActiveProfiles;
-import org.springframework.transaction.annotation.Transactional;
-
-@Target(ElementType.TYPE)
-@Retention(RetentionPolicy.RUNTIME)
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-@ActiveProfiles("test")
-@Import({BaseIntegrationTestConfig.class, WebTestConfig.class})
-@Transactional
-public @interface IntegrationTest {
-}
