--- backend/test-support/src/main/java/coffeeshout/support/config/WebTestConfig.java
@@ -1,16 +1,0 @@
-package coffeeshout.support.config;
-
-import coffeeshout.web.exception.RestExceptionHandler;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
-import org.springframework.boot.test.context.TestConfiguration;
-import org.springframework.context.annotation.Bean;
-
-@TestConfiguration(proxyBeanMethods = false)
-public class WebTestConfig {
-
-    @Bean
-    @ConditionalOnMissingBean
-    public RestExceptionHandler restExceptionHandler() {
-        return new RestExceptionHandler();
-    }
-}
